move diagnostics to domain
