+++ app/domain/diagnostic/cz/check.go
@@ -12,7 +12,7 @@ import (
 	"github.com/sirupsen/logrus"
 
 	config "github.com/cloudzero/cloudzero-agent/app/config/validator"
-	"github.com/cloudzero/cloudzero-agent/app/diagnostic"
+	"github.com/cloudzero/cloudzero-agent/app/domain/diagnostic"
 	http "github.com/cloudzero/cloudzero-agent/app/http/client"
 	logging "github.com/cloudzero/cloudzero-agent/app/logging/validator"
 	"github.com/cloudzero/cloudzero-agent/app/types/status"
