+++ app/domain/diagnostic/runner/runner_test.go
@@ -10,9 +10,9 @@ import (
 	"testing"
 
 	config "github.com/cloudzero/cloudzero-agent/app/config/validator"
-	"github.com/cloudzero/cloudzero-agent/app/diagnostic"
-	"github.com/cloudzero/cloudzero-agent/app/diagnostic/catalog"
-	"github.com/cloudzero/cloudzero-agent/app/diagnostic/kms"
+	"github.com/cloudzero/cloudzero-agent/app/domain/diagnostic"
+	"github.com/cloudzero/cloudzero-agent/app/domain/diagnostic/catalog"
+	"github.com/cloudzero/cloudzero-agent/app/domain/diagnostic/kms"
 	"github.com/cloudzero/cloudzero-agent/app/types/status"
 	"github.com/stretchr/testify/assert"
 	"k8s.io/client-go/kubernetes"
