+++ app/domain/diagnostic/stage/stage.go
@@ -9,7 +9,7 @@ import (
 	net "net/http"
 
 	config "github.com/cloudzero/cloudzero-agent/app/config/validator"
-	"github.com/cloudzero/cloudzero-agent/app/diagnostic"
+	"github.com/cloudzero/cloudzero-agent/app/domain/diagnostic"
 	logging "github.com/cloudzero/cloudzero-agent/app/logging/validator"
 	"github.com/cloudzero/cloudzero-agent/app/types/status"
 	"github.com/sirupsen/logrus"
